refactor: session doc is the handle — chat/resume live on AISession, Session wrapper removed
